+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - TDB
+### Added
+  * `Entity::getEntitiesByClass(string $name) : array`
+
 ## [0.1.0] - 2016-12-05
 ### Added
   * Initial version